+++ PSGSuite/Public/Groups/Get-GSGroupSettings.ps1
@@ -7,7 +7,7 @@ function Get-GSGroupSettings {
     Gets a group's settings
 
     .PARAMETER Identity
-    The email of the group
+    The email or unique ID of the group.
 
     If only the email name-part is passed, the full email will be contstructed using the Domain from the active config
 
@@ -35,8 +35,10 @@ function Get-GSGroupSettings {
     Process {
         try {
             foreach ($G in $Identity) {
-                if ($G -notlike "*@*.*") {
-                    $G = "$($G)@$($Script:PSGSuite.Domain)"
+                Resolve-Email ([ref]$G) -IsGroup
+                if ($G -notmatch '^[\w.%+-]+@[a-zA-Z\d.-]+\.[a-zA-Z]{2,}$') {
+                    Write-Verbose "Getting Group Email for ID '$G' as the Group Settings API only accepts Group Email addresses."
+                    $G = Get-GSGroup -Identity $G -Verbose:$false | Select-Object -ExpandProperty Email
                 }
                 Write-Verbose "Getting settings for group '$G'"
                 $request = $service.Groups.Get($G)