+++ PSGSuite/Public/Groups/Remove-GSGroupAlias.ps1
@@ -2,29 +2,27 @@ function Remove-GSGroupAlias {
     <#
     .SYNOPSIS
     Removes an alias from a G Suite group
-    
+
     .DESCRIPTION
     Removes an alias from a G Suite group
-    
-    .PARAMETER Group
+
+    .PARAMETER Identity
     The group to remove the alias from
-    
+
     .PARAMETER Alias
     The alias or list of aliases to remove from the group
-    
+
     .EXAMPLE
-    Remove-GSGroupAlias -Group humanresources@domain.com -Alias 'hr@domain.com','hrhelp@domain.com'
+    Remove-GSGroupAlias -Identity humanresources@domain.com -Alias 'hr@domain.com','hrhelp@domain.com'
 
     Removes 2 aliases for group Human Resources: 'hr@domain.com' and 'hrhelp@domain.com'
     #>
     [cmdletbinding(SupportsShouldProcess = $true,ConfirmImpact = "High")]
-    Param
-    (
-        [parameter(Mandatory = $true,Position = 0,ValueFromPipelineByPropertyName = $true)]
-        [Alias("Email")]
-        [ValidateNotNullOrEmpty()]
+    Param (
+        [parameter(Mandatory = $true,Position = 0,ValueFromPipeline = $true,ValueFromPipelineByPropertyName = $true)]
+        [Alias('GroupEmail','Group','Email')]
         [String]
-        $Group,
+        $Identity,
         [parameter(Mandatory = $true,Position = 1)]
         [String[]]
         $Alias
@@ -37,19 +35,15 @@ function Remove-GSGroupAlias {
         $service = New-GoogleService @serviceParams
     }
     Process {
+        Resolve-Email ([ref]$Identity) -IsGroup
         foreach ($A in $Alias) {
             try {
-                if ($Group -notlike "*@*.*") {
-                    $Group = "$($Group)@$($Script:PSGSuite.Domain)"
-                }
-                if ($A -notlike "*@*.*") {
-                    $A = "$($A)@$($Script:PSGSuite.Domain)"
-                }
-                if ($PSCmdlet.ShouldProcess("Removing alias '$A' from Group '$Group'")) {
-                    Write-Verbose "Removing alias '$A' from Group '$Group'"
-                    $request = $service.Groups.Aliases.Delete($Group,$A)
+                Resolve-Email ([ref]$A)
+                if ($PSCmdlet.ShouldProcess("Removing alias '$A' from Group '$Identity'")) {
+                    Write-Verbose "Removing alias '$A' from Group '$Identity'"
+                    $request = $service.Groups.Aliases.Delete($Identity,$A)
                     $request.Execute()
-                    Write-Verbose "Alias '$A' has been successfully deleted from Group '$Group'"
+                    Write-Verbose "Alias '$A' has been successfully deleted from Group '$Identity'"
                 }
             }
             catch {